pyproject.toml
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "flama"
version = "0.7.0"
description = "Fire up your API with this flamethrower"
authors = ["José Antonio Perdiguero López <perdy@perdy.io>"]
license = "GPL-3.0+"
readme = "README.md"
repository = "https://github.com/PeRDy/flama"
keywords = ["starlette", "api", "apistar", "components", "schema"]
include = []
exclude = []

[tool.poetry.dependencies]
python = "^3.6"
starlette = ">=0.11"
marshmallow = {version = "^3.0 || ^3.0.0rc5",allows-prereleases = true}
python-forge = {version = "^18.6",optional = true}
apispec = {version = "^0.39.0",optional = true}
pyyaml = {version = "^3.13",optional = true}
sqlalchemy = {version = "^1.2",optional = true}
databases = {version = "^0.1.9",optional = true}

[tool.poetry.dev-dependencies]
pytest = "^3.6"
pytest-xdist = "^1.22"
pytest-cov = "^2.5"
tox = "^3.0"
clinner = "^1.10"
ipython = "^6.4"
requests = "^2.19"
ipdb = "^0.11.0"
isort = "^4.3"
python-multipart = "^0.0.5"
pyyaml = "^3.13"
apispec = "^0.39.0"
flake8 = "^3.7"
databases = {version = "^0.1.9",extras = ["sqlite"]}
pytest-asyncio = "^0.10.0"
sqlalchemy-utils = "^0.33.11"
python-forge = "^18.6"
mkdocs = "^1.0"
mkdocs-material = "^4.0"
black = {version = "^19.3b0",allows-prereleases = true}

[tool.poetry.extras]
full = ["apispec","pyyaml","python-forge","sqlalchemy","databases"]

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.tox
  | build
  | dist
)/
'''
